backend/requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
requests==2.31.0
numpy==2.2.0
ollama==0.1.6
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
pydantic==2.9.2
python-dotenv==1.0.0